break up estimation
