+++ src/aspire/basis/__init__.py
@@ -1,12 +1,11 @@
 from .basis import Basis
-
 from .dirac import DiracBasis
 from .fb_2d import FBBasis2D
 from .fb_3d import FBBasis3D
 from .ffb_2d import FFBBasis2D
 from .ffb_3d import FFBBasis3D
+from .fpswf_2d import FPSWFBasis2D
+from .fpswf_3d import FPSWFBasis3D
 from .polar_2d import PolarBasis2D
 from .pswf_2d import PSWFBasis2D
 from .pswf_3d import PSWFBasis3D
-from .fpswf_2d import FPSWFBasis2D
-from .fpswf_3d import FPSWFBasis3D
