+++ src/aspire/basis/ffb_3d.py
@@ -3,8 +3,8 @@ import logging
 import numpy as np
 from numpy import pi
 
-from aspire.basis.basis_utils import lgwt, norm_assoc_legendre, sph_bessel
 from aspire.basis import FBBasis3D
+from aspire.basis.basis_utils import lgwt, norm_assoc_legendre, sph_bessel
 from aspire.nufft import anufft, nufft
 from aspire.utils.matlab_compat import m_flatten, m_reshape
