+++ src/aspire/basis/fpswf_3d.py
@@ -1,4 +1,4 @@
-from aspire.basis import PSWFBasis3D
+from aspire.basis.pswf_3d import PSWFBasis3D
 
 
 class FPSWFBasis3D(PSWFBasis3D):
