+++ src/aspire/commands/cov3d.py
@@ -3,9 +3,9 @@ import logging
 import click
 
 from aspire.basis import FBBasis3D
-from aspire.estimation.covar import CovarianceEstimator
-from aspire.estimation.mean import MeanEstimator
-from aspire.estimation.noise import WhiteNoiseEstimator
+from aspire.covariance import CovarianceEstimator
+from aspire.noise import WhiteNoiseEstimator
+from aspire.reconstruction import MeanEstimator
 from aspire.source.relion import RelionSource
 
 logger = logging.getLogger(__name__)
