+++ src/aspire/covariance/covar2d.py
