+++ src/aspire/denoising/denoiser_cov2d.py
@@ -1,9 +1,9 @@
 import logging
 
 from aspire.basis import FFBBasis2D
+from aspire.covariance import BatchedRotCov2D
 from aspire.denoising import Denoiser
 from aspire.denoising.denoised_src import DenoisedImageSource
-from aspire.estimation.covar2d import BatchedRotCov2D
 from aspire.utils.optimize import fill_struct
 
 logger = logging.getLogger(__name__)
