+++ src/aspire/noise/__init__.py
@@ -1,0 +1,1 @@
+from .noise import AnisotropicNoiseEstimator, NoiseEstimator, WhiteNoiseEstimator
