+++ tests/test_anisotropic_noise.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from aspire.estimation.noise import AnisotropicNoiseEstimator
+from aspire.noise import AnisotropicNoiseEstimator
 from aspire.source.simulation import Simulation
 from aspire.utils.filters import RadialCTFFilter
