+++ tutorials/examples/cov2d_simulation.py
@@ -13,7 +13,7 @@ import mrcfile
 import numpy as np
 
 from aspire.basis import FFBBasis2D
-from aspire.estimation.covar2d import RotCov2D
+from aspire.covariance import RotCov2D
 from aspire.source.simulation import Simulation
 from aspire.utils.filters import RadialCTFFilter, ScalarFilter
 from aspire.utils.matrix import anorm
