+++ tutorials/examples/image_expansion.py
@@ -10,10 +10,7 @@ import timeit
 import matplotlib.pyplot as plt
 import numpy as np
 
-from aspire.basis import FBBasis2D
-from aspire.basis import FFBBasis2D
-from aspire.basis import FPSWFBasis2D
-from aspire.basis import PSWFBasis2D
+from aspire.basis import FBBasis2D, FFBBasis2D, FPSWFBasis2D, PSWFBasis2D
 from aspire.utils.matrix import anorm
 
 logger = logging.getLogger(__name__)
