+++ tutorials/examples/preprocess_imgs_exp.py.dontrun
@@ -6,7 +6,7 @@ before starting the pipeline of reconstructing 3D map.
 
 import matplotlib.pyplot as plt
 
-from aspire.estimation.noise import WhiteNoiseEstimator
+from aspire.noise import WhiteNoiseEstimator
 from aspire.source.relion import RelionSource
 
 # Set input path and files and initialize other parameters
