+++ tutorials/examples/preprocess_imgs_sim.py
@@ -9,7 +9,7 @@ import matplotlib.pyplot as plt
 import mrcfile
 import numpy as np
 
-from aspire.estimation.noise import WhiteNoiseEstimator
+from aspire.noise import WhiteNoiseEstimator
 from aspire.source.simulation import Simulation
 from aspire.utils.filters import RadialCTFFilter, ScalarFilter
 from aspire.volume import Volume
